Implement derivative for new splines
